Handle eth_sendTransaction for SA
