+++ sample/wallet/src/main/kotlin/com/reown/sample/wallet/domain/EthSigner.kt
@@ -1,11 +1,16 @@
 package com.reown.sample.wallet.domain
 
+import com.reown.android.cacao.signature.SignatureType
+import com.reown.android.utils.cacao.sign
 import com.reown.util.bytesToHex
 import com.reown.util.hexToBytes
+import com.reown.walletkit.utils.CacaoSigner
 import org.web3j.crypto.ECKeyPair
 import org.web3j.crypto.Sign
 
 object EthSigner {
+
+    fun personalSign(message: String): String = CacaoSigner.sign(message, EthAccountDelegate.privateKey.hexToBytes(), SignatureType.EIP191).s
 
     fun signHash(hashToSign: String, privateKey: String): String {
         val dataToSign: ByteArray = if (hashToSign.startsWith("0x")) {
