+++ sample/wallet/src/main/kotlin/com/reown/sample/wallet/ui/routes/dialog_routes/session_request/SessionRequestViewModel.kt
@@ -6,34 +6,56 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.google.firebase.crashlytics.ktx.crashlytics
 import com.google.firebase.ktx.Firebase
-import com.reown.android.cacao.signature.SignatureType
 import com.reown.android.internal.common.exception.NoConnectivityException
 import com.reown.android.utils.cacao.sign
-import com.reown.sample.common.Chains
-import com.reown.sample.wallet.domain.EthAccountDelegate
-import com.reown.sample.wallet.domain.EthSigner
+import com.reown.sample.wallet.domain.Signer
+import com.reown.sample.wallet.domain.Signer.PERSONAL_SIGN_METHOD
 import com.reown.sample.wallet.domain.WCDelegate
 import com.reown.sample.wallet.ui.common.peer.PeerUI
 import com.reown.sample.wallet.ui.common.peer.toPeerUI
-import com.reown.util.hexToBytes
 import com.reown.walletkit.client.Wallet
 import com.reown.walletkit.client.WalletKit
-import com.reown.walletkit.utils.CacaoSigner
 import kotlinx.coroutines.launch
 import org.json.JSONArray
 import org.web3j.utils.Numeric.hexStringToByteArray
-import kotlin.coroutines.resume
-import kotlin.coroutines.suspendCoroutine
-
-//import uniffi.uniffi_yttrium.Transaction
 
 class SessionRequestViewModel : ViewModel() {
     var sessionRequestUI: SessionRequestUI = generateSessionRequestUI()
 
-    private fun clearSessionRequest() {
-        WCDelegate.sessionRequestEvent = null
-        WCDelegate.currentId = null
-        sessionRequestUI = SessionRequestUI.Initial
+    fun approve(onSuccess: (Uri?) -> Unit = {}, onError: (Throwable) -> Unit = {}) {
+        viewModelScope.launch {
+            try {
+                val sessionRequest = sessionRequestUI as? SessionRequestUI.Content
+                if (sessionRequest != null) {
+                    val result: String = Signer.sign(sessionRequest)
+                    println("kobe: Result: $result")
+                    val response = Wallet.Params.SessionRequestResponse(
+                        sessionTopic = sessionRequest.topic,
+                        jsonRpcResponse = Wallet.Model.JsonRpcResponse.JsonRpcResult(sessionRequest.requestId, result)
+                    )
+                    val redirect = WalletKit.getActiveSessionByTopic(sessionRequest.topic)?.redirect?.toUri()
+                    WalletKit.respondSessionRequest(response,
+                        onSuccess = {
+                            clearSessionRequest()
+                            onSuccess(redirect)
+                        },
+                        onError = { error ->
+                            Firebase.crashlytics.recordException(error.throwable)
+                            if (error.throwable !is NoConnectivityException) {
+                                clearSessionRequest()
+                            }
+                            onError(error.throwable)
+                        })
+                } else {
+                    onError(Throwable("Approve - Cannot find session request"))
+                }
+            } catch (e: Exception) {
+                Firebase.crashlytics.recordException(e)
+                reject()
+                clearSessionRequest()
+                onError(Throwable(e.message ?: "Undefined error, please check your Internet connection"))
+            }
+        }
     }
 
     fun reject(onSuccess: (Uri?) -> Unit = {}, onError: (Throwable) -> Unit = {}) {
@@ -67,7 +89,7 @@ class SessionRequestViewModel : ViewModel() {
         } catch (e: Exception) {
             Firebase.crashlytics.recordException(e)
             clearSessionRequest()
-            onError(e.cause ?: Throwable("Undefined error, please check your Internet connection"))
+            onError(Throwable(e.message ?: "Undefined error, please check your Internet connection"))
         }
     }
 
@@ -84,86 +106,10 @@ class SessionRequestViewModel : ViewModel() {
         }
     }
 
-    fun approve(onSuccess: (Uri?) -> Unit = {}, onError: (Throwable) -> Unit = {}) {
-        viewModelScope.launch {
-            try {
-                val sessionRequest = sessionRequestUI as? SessionRequestUI.Content
-                if (sessionRequest != null) {
-                    val result: String = when {
-                        sessionRequest.method == "wallet_sendCalls" -> {
-                            val transactions: MutableList<Wallet.Model.Transaction> = mutableListOf()
-                            println("kobe: wallet_sendCalls: ${sessionRequest.param}")
-                            val callsArray = JSONArray(sessionRequest.param).getJSONObject(0).getJSONArray("calls")
-
-                            for (i in 0 until callsArray.length()) {
-                                val call = callsArray.getJSONObject(i)
-                                val to = call.getString("to") ?: ""
-                                val value = call.getString("value") ?: ""
-                                val data = call.getString("data") ?: ""
-                                transactions.add(Wallet.Model.Transaction(to, value, data))
-                            }
-                            println("kobe: Transactions: $transactions")
-
-                            val ownerAccount = Wallet.Model.Account(EthAccountDelegate.sepoliaAddress)
-                            val prepareSendTxsParams = Wallet.Params.PrepareSendTransactions(transactions = transactions, owner = ownerAccount)
-
-                            suspendCoroutine<String> { continuation ->
-                                WalletKit.prepareSendTransactions(prepareSendTxsParams) { result ->
-
-                                    val signature = EthSigner.signHash(result.hash, EthAccountDelegate.privateKey)
-                                    val doSendTxsParams = Wallet.Params.DoSendTransactions(
-                                        owner = ownerAccount,
-                                        signatures = listOf(Wallet.Model.OwnerSignature(address = EthAccountDelegate.account, signature = signature)),
-                                        doSendTransactionParams = result.doSendTransactionParams
-                                    )
-
-                                    WalletKit.doSendTransactions(doSendTxsParams) { doTxsResult -> continuation.resume(doTxsResult.userOperationHash) }
-                                }
-                            }
-                        }
-
-                        sessionRequest.method == PERSONAL_SIGN_METHOD -> CacaoSigner.sign(sessionRequest.param, EthAccountDelegate.privateKey.hexToBytes(), SignatureType.EIP191).s
-                        sessionRequest.chain?.contains(Chains.Info.Eth.chain, true) == true ->
-                            """0xa3f20717a250c2b0b729b7e5becbff67fdaef7e0699da4de7ca5895b02a170a12d887fd3b17bfdce3481f10bea41f45ba9f709d39ce8325427b57afcfc994cee1b"""
-
-                        sessionRequest.chain?.contains(Chains.Info.Cosmos.chain, true) == true ->
-                            """{"signature":"pBvp1bMiX6GiWmfYmkFmfcZdekJc19GbZQanqaGa\/kLPWjoYjaJWYttvm17WoDMyn4oROas4JLu5oKQVRIj911==","pub_key":{"value":"psclI0DNfWq6cOlGrKD9wNXPxbUsng6Fei77XjwdkPSt","type":"tendermint\/PubKeySecp256k1"}}"""
-                        //Note: Only for testing purposes - it will always fail on Dapp side
-                        sessionRequest.chain?.contains(Chains.Info.Solana.chain, true) == true ->
-                            """{"signature":"pBvp1bMiX6GiWmfYmkFmfcZdekJc19GbZQanqaGa\/kLPWjoYjaJWYttvm17WoDMyn4oROas4JLu5oKQVRIj911==","pub_key":{"value":"psclI0DNfWq6cOlGrKD9wNXPxbUsng6Fei77XjwdkPSt","type":"tendermint\/PubKeySecp256k1"}}"""
-
-                        else -> throw Exception("Unsupported Chain")
-                    }
-
-
-                    println("kobe: Result: $result")
-                    val response = Wallet.Params.SessionRequestResponse(
-                        sessionTopic = sessionRequest.topic,
-                        jsonRpcResponse = Wallet.Model.JsonRpcResponse.JsonRpcResult(sessionRequest.requestId, result)
-                    )
-
-                    val redirect = WalletKit.getActiveSessionByTopic(sessionRequest.topic)?.redirect?.toUri()
-                    WalletKit.respondSessionRequest(response,
-                        onSuccess = {
-                            clearSessionRequest()
-                            onSuccess(redirect)
-                        },
-                        onError = { error ->
-                            Firebase.crashlytics.recordException(error.throwable)
-                            if (error.throwable !is NoConnectivityException) {
-                                clearSessionRequest()
-                            }
-                            onError(error.throwable)
-                        })
-                } else {
-                    onError(Throwable("Approve - Cannot find session request"))
-                }
-            } catch (e: Exception) {
-                Firebase.crashlytics.recordException(e)
-                clearSessionRequest()
-                onError(e.cause ?: Throwable("Undefined error, please check your Internet connection"))
-            }
-        }
+    private fun clearSessionRequest() {
+        WCDelegate.sessionRequestEvent = null
+        WCDelegate.currentId = null
+        sessionRequestUI = SessionRequestUI.Initial
     }
 
     private fun generateSessionRequestUI(): SessionRequestUI {
@@ -189,5 +135,3 @@ class SessionRequestViewModel : ViewModel() {
         }
     }
 }
-
-private const val PERSONAL_SIGN_METHOD = "personal_sign"
